Refactor search POJOs to avoid conflicts with collection files.
