+++ app/src/main/java/org/metabrainz/mobile/data/repository/SearchRepository.java
@@ -12,13 +12,13 @@ import org.metabrainz.mobile.data.sources.api.entities.mbentity.Label;
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.Recording;
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.Release;
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.ReleaseGroup;
-import org.metabrainz.mobile.data.sources.api.entities.response.ArtistSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.EventSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.InstrumentSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.LabelSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.RecordingSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.ReleaseGroupSearchResponse;
-import org.metabrainz.mobile.data.sources.api.entities.response.ReleaseSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.ArtistSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.EventSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.InstrumentSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.LabelSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.RecordingSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.ReleaseGroupSearchResponse;
+import org.metabrainz.mobile.data.sources.api.entities.response.search.ReleaseSearchResponse;
 import org.metabrainz.mobile.util.Log;
 
 import java.util.List;
