+++ app/src/main/java/org/metabrainz/mobile/data/sources/api/entities/response/search/EventSearchResponse.java
@@ -1,10 +1,10 @@
-package org.metabrainz.mobile.data.sources.api.entities.response;
+package org.metabrainz.mobile.data.sources.api.entities.response.search;
 
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.Event;
 
 import java.util.List;
 
-public class EventSearchResponse extends JSONResponse {
+public class EventSearchResponse extends JSONSearchResponse {
     private List<Event> events;
 
     public List<Event> getEvents() {
