+++ app/src/main/java/org/metabrainz/mobile/data/sources/api/entities/response/search/InstrumentSearchResponse.java
@@ -1,11 +1,11 @@
-package org.metabrainz.mobile.data.sources.api.entities.response;
+package org.metabrainz.mobile.data.sources.api.entities.response.search;
 
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.Instrument;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class InstrumentSearchResponse extends JSONResponse {
+public class InstrumentSearchResponse extends JSONSearchResponse {
     private List<Instrument> instruments = new ArrayList<>();
 
     public List<Instrument> getInstruments() {
