+++ app/src/main/java/org/metabrainz/mobile/data/sources/api/entities/response/search/RecordingSearchResponse.java
@@ -1,10 +1,10 @@
-package org.metabrainz.mobile.data.sources.api.entities.response;
+package org.metabrainz.mobile.data.sources.api.entities.response.search;
 
 import org.metabrainz.mobile.data.sources.api.entities.mbentity.Recording;
 
 import java.util.List;
 
-public class RecordingSearchResponse extends JSONResponse {
+public class RecordingSearchResponse extends JSONSearchResponse {
 
     private List<Recording> recordings;
