+++ app/src/main/java/org/metabrainz/mobile/data/sources/api/entities/response/search/ReleaseGroupSearchResponse.java
@@ -1,4 +1,4 @@
-package org.metabrainz.mobile.data.sources.api.entities.response;
+package org.metabrainz.mobile.data.sources.api.entities.response.search;
 
 import com.google.gson.annotations.SerializedName;
 
@@ -7,7 +7,7 @@ import org.metabrainz.mobile.data.sources.api.entities.mbentity.ReleaseGroup;
 import java.util.ArrayList;
 import java.util.List;
 
-public class ReleaseGroupSearchResponse extends JSONResponse {
+public class ReleaseGroupSearchResponse extends JSONSearchResponse {
     @SerializedName("release-groups")
     private List<ReleaseGroup> releaseGroups = new ArrayList<>();
